pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.86.0",
    "pillow>=11.2.1",
    "pyaudio>=0.2.14",
    "requests>=2.32.4",
    "speechrecognition>=3.14.3",
    "streamlit>=1.45.1",
]
